Initial commit
So, this project is to show the application of tries data structure as well as heap data structure.
As you see in your web search , we have autocomplete. This can be done using tries . We can also observe that our search history is shown on the top among all the suggestions . We try to show how this can be achieved using a history(browse.txt) file and a update file .